# swagger.yaml
swagger: '2.0'
info:
  title: SE2
  description: SE2 is a simple API allowing user to create exam and tasks.
  version: "1"
host: se2-2018.herokuapp.com
basePath: /v1
schemes:
  - https
produces:
  - application/json
paths:
...
  /TA-peer-corrections/{TAPeerCorrectionID}:
    parameters:
      - in: path
        name: TAPeerCorrectionID
        required: true
        type: integer
    get:
      tags:
      - "TA peer corrections"
      summary: Returns a TA-peer-correction by ID.
      responses:
        200:
          description: The requested TA-peer-correction. 
          schema:
            $ref: '#/definitions/TA-peer-correction'
        404:
          description: Not Found
    put:
      tags:
      - "TA peer corrections"
      summary: Update a TA-peer-correction by ID.
      parameters:
        - in: body
          name: TA-peer-correction
          description: The TA-peer-correction to update
          required: true
          schema:
            $ref: '#/definitions/TA-peer-correction_no_id'
      responses:
        200:
          description: OK, updated.
        409:
          description: Conflict
    delete:
      tags:
      - "TA peer corrections"
      summary: Delete a TA-peer-correction by ID.
      responses:
        204:
          description: OK, deleted.
        404:
          description: Not Found
...
definitions:
...
  TA-peer-correction:
    type: "object"
    properties:
      id:
        description: The ID of the TA-peer-correction
        type: integer
      peer-correction_id:
        description: The ID of the peer-correction which the TA-peer-correction refers to
        type: integer
      text:
        description: The text of the TA-peer-correction
        type: string
      score:
        description: The score of the TA-peer-correction
        type: integer
      user_id:
        description: The ID of the TA who is correcting the peer-correction of a student
        type: integer
    required:
      - id
      - peer-correction_id
      - text
      - score
      - user_id
  TA-peer-correction_no_id:
    type: "object"
    properties:
      peer-correction_id:
        description: The ID of the peer-correction which the TA-peer-correction refers to
        type: integer
      text:
        description: The text of the TA-peer-correction
        type: string
      score:
        description: The score of the TA-peer-correction
        type: integer
      user_id:
        description: The ID of the TA who is correcting the peer-correction of a student
        type: integer
    required:
      - peer-correction_id
      - text
      - score
      - user_id
...